[WIP] 21-eteams-admin-menu

+++ app/Http/Managers/EteamMemberManager.php
@@ -6,6 +6,9 @@ namespace App\Http\Managers;
 
 use App\Http\Repositories\EteamMemberRepository;
 use App\Http\Services\EteamMemberService;
+use App\Models\ETeamLog;
+use App\Models\User;
+use Illuminate\Support\Str;
 
 class EteamMemberManager
 {
@@ -54,15 +57,18 @@ class EteamMemberManager
 
     public const REG_NOT_EXISTS = 'El usuario ya no pertenece al equipo';
 
-    private EteamMemberService $eteamMemberService;
     private EteamMemberRepository $eteamMemberRepository;
+    private EteamPostManager $eteamPostManager;
+    private EteamLogManager $eteamLogManager;
 
     public function __construct(
-        EteamMemberService $eteamMemberService,
-        EteamMemberRepository $eteamMemberRepository
+        EteamMemberRepository $eteamMemberRepository,
+        EteamPostManager $eteamPostManager,
+        EteamLogManager $eteamLogManager
     ) {
-        $this->eteamMemberService = $eteamMemberService;
         $this->eteamMemberRepository = $eteamMemberRepository;
+        $this->eteamPostManager = $eteamPostManager;
+        $this->eteamLogManager = $eteamLogManager;
     }
 
     public function transferTeamOwnership(int $eteamId, int $oldOwnerId, int $newOwnerId): bool
@@ -70,13 +76,50 @@ class EteamMemberManager
         return $this->eteamMemberRepository->transferTeamOwnership($eteamId, $oldOwnerId, $newOwnerId);
     }
 
-    public function grantCaptainRange(int $eteamId, int $memberId): bool
+    public function updateCaptainRange(int $eteamId, int $memberId, string $range): void
     {
-        return $this->eteamMemberRepository->grantCaptainRange($eteamId, $memberId);
+        //        $admin =
+        $user = User::find($memberId);
+        $logMessage = '';
+
+        if ($range === 'captain') {
+            $this->eteamMemberRepository->grantCaptainRange($eteamId, $memberId);
+            $logMessage = $user->name . ' cambia su rango en el equipo, miembro -> capitán';
+            $postTitle = $user->name. ' asciende a capitán';
+            $postContent = '';
+        }
+
+        if ($range === 'member') {
+            $this->eteamMemberRepository->removeCaptainRange($eteamId, $memberId);
+            $logMessage = $user->name . ' cambia su rango en el equipo, capitán -> miembro';
+            $postTitle = $user->name. ' deja la capitanía del equipo';
+            $postContent = '';
+        }
+
+        $logData = [
+            'eteam_id' => $eteamId,
+            'user_id' => $memberId,
+            'context' => ETeamLog::CONTEXT_MEMBERS,
+            'type' => ETeamLog::TYPE_UPDATE,
+            'message' => $logMessage
+        ];
+        $this->eteamLogManager->create($logData);
+
+        $postData = [
+            'eteam_id' => $eteamId,
+            'user_id' => $memberId,
+            'title' => $postTitle,
+            'content' => $postContent,
+            'public' => false,
+            'slug' => Str::slug($postTitle, '-')
+        ];
+        $this->eteamPostManager->create($postData);
+
+        // enviar personal al usuario implicado y mail a todos los miembros
     }
 
-    public function removeCaptainRange(int $eteamId, int $memberId): bool
+    public function removeCaptainRange(int $eteamId, int $memberId): void
     {
-        return $this->eteamMemberRepository->removeCaptainRange($eteamId, $memberId);
+        $this->eteamMemberRepository->removeCaptainRange($eteamId, $memberId);
     }
 }
